fix(ack deploy): Fix deployment failure issue

+++ boost.common/src/main/java/org/example/common/param/commodity/specification/GetCommodityPriceParam.java
@@ -24,7 +24,7 @@ public class GetCommodityPriceParam extends SpiBaseParam {
     /**
      * 支付周期
      */
-    private Integer payPeriod;
+    private Long payPeriod;
 
     /**
      * 支付周期单位
